task: do not remove extensions
Do not remove extensions from "task names".

Removing the extensions kind-of-makes sense when the task name is a
filename, like `a.rkt`, but it definitely does not work when the task
name is a directory name that contains a `.`.

So just use the full name for the `*.in` and `*.out` files